requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
Levenshtein==0.21.1
MonthDelta==0.9.1
multidict==6.0.4
py-cord==2.6.1
python-dotenv==0.20.0
python-Levenshtein==0.21.1
rapidfuzz==3.1.2
thefuzz==0.19.0
yarl==1.9.2
